Moved url bar related files into app folder
Resolves #8188

Auditors: @bridiver

Test Plan:
- all test should be green
