+++ app/renderer/components/navigation/navigationBar.js
@@ -4,27 +4,27 @@
 
 const React = require('react')
 const Immutable = require('immutable')
-const ImmutableComponent = require('./immutableComponent')
+const ImmutableComponent = require('./../../../../js/components/immutableComponent')
 
-const cx = require('../lib/classSet')
-const Button = require('./button')
-const UrlBar = require('../../app/renderer/components/urlBar')
-const windowActions = require('../actions/windowActions')
-const appActions = require('../actions/appActions')
-const siteTags = require('../constants/siteTags')
-const messages = require('../constants/messages')
-const settings = require('../constants/settings')
+const cx = require('../../../../js/lib/classSet')
+const Button = require('./../../../../js/components/button')
+const UrlBar = require('./urlBar')
+const windowActions = require('../../../../js/actions/windowActions')
+const appActions = require('../../../../js/actions/appActions')
+const siteTags = require('../../../../js/constants/siteTags')
+const messages = require('../../../../js/constants/messages')
+const settings = require('../../../../js/constants/settings')
 const ipc = require('electron').ipcRenderer
-const {isSourceAboutUrl, getBaseUrl} = require('../lib/appUrlUtil')
-const AddEditBookmarkHanger = require('../../app/renderer/components/addEditBookmarkHanger')
-const siteUtil = require('../state/siteUtil')
-const eventUtil = require('../lib/eventUtil')
-const UrlUtil = require('../lib/urlutil')
-const getSetting = require('../settings').getSetting
-const windowStore = require('../stores/windowStore')
-const contextMenus = require('../contextMenus')
-const LongPressButton = require('./longPressButton')
-const PublisherToggle = require('../../app/renderer/components/publisherToggle')
+const {isSourceAboutUrl, getBaseUrl} = require('../../../../js/lib/appUrlUtil')
+const AddEditBookmarkHanger = require('../addEditBookmarkHanger')
+const siteUtil = require('../../../../js/state/siteUtil')
+const eventUtil = require('../../../../js/lib/eventUtil')
+const UrlUtil = require('../../../../js/lib/urlutil')
+const getSetting = require('../../../../js/settings').getSetting
+const windowStore = require('../../../../js/stores/windowStore')
+const contextMenus = require('../../../../js/contextMenus')
+const LongPressButton = require('./../../../../js/components/longPressButton')
+const PublisherToggle = require('../publisherToggle')
 
 class NavigationBar extends ImmutableComponent {
   constructor () {
